REFACTOR: Update import paths for prisma to db.client in index, auth.service, and user.service files
